fix: correct types & interfaces structure
