--- src/components/Docs/types/types.ts
@@ -1,3 +1,0 @@
-export const isNonNull = <T>(value: T): boolean => {
-  return value !== null && value !== undefined;
-};
