+++ src/types/interface.ts
@@ -1,10 +1,12 @@
 import {
+  GraphQLField,
   GraphQLFieldMap,
   GraphQLObjectType,
   GraphQLOutputType,
+  GraphQLType,
   IntrospectionQuery,
 } from 'graphql';
-import { RootFieldType } from './types';
+import { Maybe, RootFieldType } from './types';
 
 export interface DataForm extends DataLogin {
   confirmPassword: string;
@@ -61,4 +63,15 @@ export interface CustomSectionProps {
   title: string;
   attentionTranslation: string;
   initialActionValue: string;
+}
+
+export interface IFieldSchema {
+  schema: Maybe<IRootSchema>;
+  setRootSchema: (value: IRootSchema) => void;
+}
+
+export interface INextField {
+  handleClick: (value: GraphQLType) => void;
+  fieldName: string;
+  fieldType: GraphQLOutputType | GraphQLField<unknown, unknown>;
 }
