+++ src/types/types.ts
@@ -49,3 +49,7 @@ export const isGetFieldsType = (
     value instanceof GraphQLInputObjectType
   );
 };
+
+export const isNonNull = <T>(value: T): boolean => {
+  return value !== null && value !== undefined;
+};
